+++ crates/net/network/tests/it/connect.rs
@@ -3,7 +3,6 @@
 use alloy_node_bindings::Geth;
 use alloy_provider::{ext::AdminApi, ProviderBuilder};
 use futures::StreamExt;
-use reth_chainspec::net::mainnet_nodes;
 use reth_discv4::Discv4Config;
 use reth_eth_wire::DisconnectReason;
 use reth_net_banlist::BanList;
@@ -16,7 +15,7 @@ use reth_network_p2p::{
     headers::client::{HeadersClient, HeadersRequest},
     sync::{NetworkSyncUpdater, SyncState},
 };
-use reth_network_peers::NodeRecord;
+use reth_network_peers::{mainnet_nodes, NodeRecord};
 use reth_primitives::HeadersDirection;
 use reth_provider::test_utils::NoopProvider;
 use reth_transaction_pool::test_utils::testing_pool;
